refactor(scripts-guild): make guild config optional, fix undefined guildIds
