+++ scripts/modules/configHandler.js
@@ -1,4 +1,3 @@
 import rootSettingsJson from "../../settings.json" with { type: "json" }
-import scriptSettingsJson from "../settings.json" with { type: "json" }
 
-export { rootSettingsJson, scriptSettingsJson }
+export { rootSettingsJson }
